Added a function to get the MD5 checksum of a given string or the contents of a given file.
It will be used later to implement a cache preventing fetching all classes for all types of files each
time an import is requested. For each file, before actually fetching the classes encompassed, a hash value
will be computed form the file contents and checked against the checksum value stored in the cache.
If the checksum matches, the previously fetched imports are used directly without computation. If the hash
is different then classes are fetched as usual.